% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_mice_to_groups.R
\name{identify_outliers}
\alias{identify_outliers}
\title{Identify Outliers Using the 1.5*IQR Rule}
\usage{
identify_outliers(data, tumor_col)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{tumor_col}{The name of the tumor size column (as a symbol).}
}
\value{
A list containing:
\item{data_no_outliers}{Data frame without outliers.}
\item{outliers}{Data frame with outliers.}
}
\description{
This function identifies outliers in the tumor size data.
}
